+++ releasetest.py
@@ -1,7 +1,7 @@
 import pytzer as pz
 #mols, ions, tempK, pres, prmlib, Gex_nRT, osm, aw, acfs =  pz.blackbox('testfiles/pytzerQuickStart.csv')
 
-import pytzer as pz
+
 mols, ions, tempK, pres = pz.io.getmols('testfiles/pytzerQuickStart.csv')
 # Calculate water activity
 aw_Seawater = pz.model.aw(mols, ions, tempK, pres)